Fix issue with mismatched background modes on open
When opening Emacs, if the background mode was currently incorrect for
the time of day, it could remain incorrect for the rest of the day until
a correct timer was instantiated.

This fix ensures the background mode gets corrected for the time of day,
and then proceeds to initialise the background timers as normal.