Prevent sync from hanging when IDs change
If you added a new heading to a file and tried to sync, the IDs would
get updated, but the file would not be saved. This caused the process
to hang as it asked for confirmation on whether to save the files.

Now we just ensure that the files get saved whenever the IDs are
updated.