Add configuration for =maybe.org= GTD file
Originally I just had a 'someday' file to contain all inactive ideas,
projects, etc., but this grew very large and unwieldy. This change
introduces a new 'maybe' file, which will allow easier division
between projects I wish to engage in eventually, and those for which I
am undecided, or currently (and for a long time) unable to engage
with.

This change adds configuration for the new 'maybe' file, and
additionally updates some of the descriptions for the 'someday' file
to better distinguish its purpose from the 'maybe' file.